[ownsem][licm] use ownership semantics (if enabled) for hoist/sink
This enables two kinds of optimizations:
1. Increase number of promotion load/store candidates by not bailing out
   in the presence of unknown calls unless unknown function directly
   takes in arguments.
2. Not bail out when promoting loop mem accesses to scalars in the
   presence of function calls that may have implicit control flow.

For 1, we rely on stacked borrows semantics that ensures that interleaved
writes between caller and callee is invalid for owned, borrowed pointers.

For 2,  we add an assumption that on implicit control flow (exception),
writes to borrowed and owned pointers are invisible to the rest of the
program since the function did not return normally.